fix(lint): fix clang-tidy warnings in create, merge, serial
- cmd/image/create.c: cast 1024u to size_t before multiplication to
  avoid implicit widening from unsigned int to size_t.
- cmd/image/merge.c: same 1024u fix; use a local index variable for
  the 2*i argv offset to avoid implicit int->ptrdiff_t widening;
  reject total==0 before malloc to satisfy the portability checker.
- platform/posix/serial.c: guard B57600/B115200/B230400 with #ifdef
  like the higher baud rates -- macOS hides them behind feature
  macros that _POSIX_C_SOURCE=200112L does not enable.

+++ cmd/image/create.c
@@ -154,7 +154,7 @@ static size_t flash_size_bytes(const char *s)
 		die("unrecognised flash size '%s' (expected e.g. 2MB or 4KB)",
 		    s);
 	}
-	return (size_t)mb * 1024u * 1024u;
+	return (size_t)mb * (size_t)1024u * (size_t)1024u;
 }
 
 int cmd_image_create(int argc, const char **argv)
